+++ src/submissions/frontend/widgets/omni_search.py
@@ -60,6 +60,7 @@ class SearchBox(QDialog):
         Changes form inputs based on sample type
         """
         search_fields = []
+        # search_fields = self.object_type.searchables
         logger.debug(f"Search fields: {search_fields}")
         deletes = [item for item in self.findChildren(FieldSearch)]
         for item in deletes:
@@ -68,7 +69,7 @@ class SearchBox(QDialog):
         if not self.sub_class:
             logger.warning(f"No subclass selected.")
             self.update_data()
-            return
+            # return
         else:
             if self.sub_class.currentText() == "Any":
                 self.object_type = self.original_type